Important Changes
changes to log function and othe folder structure changes
