+++ css/style_index.css
@@ -1,0 +1,76 @@
+body{background-color: linen;
+    border-spacing: 1px;
+	font-weight: 300;
+	font-family: Arial;
+}
+
+
+.title-super{
+	text-align: center;
+   font-size: 50px;
+   color:#000000;
+}
+
+.title-super2{
+   font-size: 30px;
+   color:#000000;
+}
+
+.title-super3{
+	text-align: right;
+   font-size: 30px;
+   color:#000000;
+}
+
+.title-logo{
+	margin-top:20px;
+}
+
+.login{border:2px solid black;
+	background-color: lightblue;
+	padding : 10px;
+	height: 500px;width:600px;
+	max-width: 700px;
+	font-size: 18;
+	font-family: Arial;
+}
+.sign-up{border:2px solid black;
+	padding : 10px;
+	background-color: lightblue;
+	height: 500px;width:600px;
+	max-width: 700px;
+	font-size: 18;
+	font-family: Arial;
+	
+}
+
+h5{
+	text-align: justify-all;
+	font-size: 20px;
+}
+
+.image-main{
+	width:700px;
+	height:400px;
+
+}
+
+p.log_details {
+	font-style: italic;
+	font-family: Times;
+	font-size: 18px;
+}
+
+#frm01{
+	background-color: white;
+}
+
+.choosefile{
+	align-items:  center;
+	text-align: center;
+}
+
+.submit{
+	align-items: center;
+	text-align: center;
+}
