+++ templates/experiments.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="ISO-8859-1">
+<title>Experiments</title>
+</head>
+<body>
+<a href="http://localhost:8000/uploader"><h2>Logarithmic</h2></a>
+<a href="http://localhost:8000/histequal"><h2>Histogram Equalization</h2></a>
+</body>
+</html>
