+++ templates/index.html
@@ -7,26 +7,26 @@
  <meta http-equiv="X-UA-Compatible" content="IE=edge">
  <meta name="viewport" content="width=device-width, initial-scale=1">
  <title>Index Page</title>
- <link rel="stylesheet" type="text/css" href="css/bootstrap.css">
- <link rel="stylesheet" href="style_index.css">
+ <link rel="stylesheet" type="text/css" href="../css/bootstrap.css">
+ <link rel="stylesheet" type="text/css" href="../css/style_index.css">
 </head>
 
 
 <body>
-  <div class="container">	
+  <div class="container">
  	<div class="row">
  		<div class="col-md-12 text-center">
   			<h1 class="title-super" > Virtual Image Processing Lab</h1>
   		</div>
-  	</div>		
-  	<div class="row">	
+  	</div>
+  	<div class="row">
  		<div class="col-md-6">
- 			<img class="image_logo" src="Data_webpage/logo_csre.jpg" alt="CSRE logo"></img>
+ 			<img class="image_logo" src="../images/logo_csre.jpg" alt="CSRE logo"></img>
  		</div>
  		<div class="col-md-6">
  			<h3 class="title-super3 "> IIT Bombay </h3>
- 		</div>	
- 	</div>	
+ 		</div>
+ 	</div>
 
   	<div class="row">
   		<div class="col-md-12">
@@ -35,38 +35,38 @@
   	</div>
   	<div class="row">
     	<div class="col-md-12">
-       		<br><br><br> 
-    	</div> 
+       		<br><br><br>
+    	</div>
   	</div>
 
   	<div class="row">
     	<div class="col-md-12 text-center">
-     		<img class="image-main" src="Data_webpage/index_image.jpg" alt="Image Processing">
+     		<img class="image-main" src="../images/index_image.jpg" alt="Image Processing">
     	</div>
   	</div>
 
   	<div class="row">
     	<div class="col-md-12">
-       		<br><br><br><br><br>  
-    	</div> 
+       		<br><br><br><br><br>
+    	</div>
   	</div>
 
   	<div class="row">
  		<div class="col-md-3">
-  			<a href="http://localhost:5000/about"><h2>About</h2></a>
+  			<a href="http://localhost:8000/about"><h2>About</h2></a>
   		</div>
    		<div class="col-md-3">
-   			<a href="http://localhost:5000/experiments"><h2>Experiments</h2></a>	
+   			<a href="http://localhost:8000/experiments"><h2>Experiments</h2></a>
    		</div>
    		<div class="col-md-3">
-   			<a href="http://localhost:5000/feedback"><h2>Feedback</h2></a>	
+   			<a href="http://localhost:8000/feedback"><h2>Feedback</h2></a>
    		</div>
    		<div class="col-md-3">
-   			<a href="http://localhost:5000/login"><h2>Login</h2></a>	
+   			<a href="http://localhost:8000/login"><h2>Login</h2></a>
    		</div>
    	</div>
-  </div>   
-  		
-   
+  </div>
+
+
 </body>
 </html>
