+++ templates/upload.html
@@ -7,25 +7,25 @@
  <meta http-equiv="X-UA-Compatible" content="IE=edge">
  <meta name="viewport" content="width=device-width, initial-scale=1">
  <title>Experiment 1</title>
- <link rel="stylesheet" type="text/css" href="css/bootstrap.css">
- <link rel="stylesheet" href="style_index.css">
+ <link rel="stylesheet" type="text/css" href="../css/bootstrap.css">
+ <link rel="stylesheet" href="../style_index.css">
 </head>
 
 <body>
-	<div class="container">	
+	<div class="container">
  		<div class="row">
  			<div class="col-md-12 text-center">
   				<h1 class="title-super" > Virtual Image Processing Lab</h1>
   			</div>
-  		</div>		
-  		<div class="row">	
+  		</div>
+  		<div class="row">
  			<div class="col-md-6">
  				<img class="image_logo" src="Data_webpage/logo_csre.jpg" alt="CSRE logo"></img>
  			</div>
  			<div class="col-md-6">
  				<h3 class="title-super3 "> IIT Bombay </h3>
- 			</div>	
- 		</div>	
+ 			</div>
+ 		</div>
 
  		<div class="row">
  			<div class="col-md-12">
@@ -35,8 +35,8 @@
 
  		<div class="row">
     		<div class="col-md-12">
-       			<br><br><br> 
-    		</div> 
+       			<br><br><br>
+    		</div>
   		</div>
 
  		<div class="row">
@@ -44,14 +44,14 @@
  				<frameset>
  			   		<legend><h3> Logarithmic Stretch</h3></legend>
  			  		<p class="log_details"> Blah blah.....</p>
- 			  	</frameset>	
+ 			  	</frameset>
  			</div>
- 				
+
  			<div class="col-md-2">
  			</div>
 
  			<div class="col-md-5">
- 			 <form action="http://localhost:5000/uploader" method="POST" enctype="multipart/form-data">
+ 			 <form action="http://localhost:8000/uploader" method="POST" enctype="multipart/form-data">
  			  <frameset >
  			  	<legend>
  			  		<h3>Choose an image and it's header file</h3>
@@ -62,6 +62,6 @@
  			  </frameset>
  			 </form>
  			</div>
- 		</div> 
+ 		</div>
 </body>
 </html>
